refactor : directory structure
